extract txpreinit
