+++ ydb/core/persqueue/read_balancer_txwrite.h
@@ -8,8 +8,6 @@
 namespace NKikimr {
 namespace NPQ {
 
-class TPersQueueReadBalancer;
-
 using namespace NTabletFlatExecutor;
 using namespace NPQRBPrivate;
